del : 팩토리 삭제
