--- src/class/monster/boss.monster.js
@@ -1,11 +1,0 @@
-import defaultMonster from "./default.monster.js";
-
-export class bossMonster extends defaultMonster
-{
-    constructor()
-    {
-        super();
-    }
-
-
-}
