--- src/class/monster/factory/default.monsterFactory.js
@@ -1,8 +1,0 @@
-//몬스터를 생성할 때의 추상 클래스. 이 팩토리 클래스에서 다양하게 생성할 수 있도록 한다.
-export class monsterFactory {
-    createMonster(id, level, monsterCode)
-    {
-        
-        throw new Error("monsterFactory는 추상 함수라 작동하지 않음")
-    }
-}
