--- src/class/monster/factory/miniMonsterFactory.js
@@ -1,11 +1,0 @@
-import { monsterFactory } from "./default.monsterFactory.js";
-import { miniMonster } from "../miniMonster.js";
-
-
-export class miniMonsterFactory extends monsterFactory
-{
-    createMonster(id,level, monsterCode)
-    {
-        return new miniMonster(id,level, monsterCode);
-    }
-}
