+++ src/init/update/update.monster.js
@@ -1,1 +1,1 @@
-//여기는 몬스터에 관한 코드를
+//여기는 몬스터에 관한 코드를 작업할 생각
